etl-pipeline/requirements.txt
# Core dependencies
pandas==2.1.0
numpy==1.24.3
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pyyaml==6.0.1
python-dotenv==1.0.0

# Data validation and quality
great-expectations==0.17.19
pydantic==2.3.0

# API and HTTP
requests==2.31.0
aiohttp==3.8.5

# Logging and monitoring
loguru==0.7.0
prometheus-client==0.17.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.1.0
isort==5.12.0 